Reorganize Redpiler in preparation for LLVM and Cranelift backends.
