+++ Cargo.toml
@@ -45,3 +45,7 @@ regex = "1.4.3"
 lazy_static = "1.4.0"
 backtrace = "0.3.56"
 rusqlite = { version="0.24.2", features=["bundled"] }
+# TODO: Put the cranelift stuff behind a feature flag
+cranelift = "0.73.0"
+cranelift-jit = "0.73.0"
+cranelift-module = "0.73.0"
